Added options to change boundary conditions. Added shear velocity to output.
Initialize weighting factors uniformly rather than zero.
In case bi < 0 than the grain size distribution in the air becomes important.
In the first time step this distribution cannot be determined. Therefore it falls back to the grain size distribution in the bed.
This has a major impact on the initial (implcitly) computed transport curve in the first time step, which in turn has a major influence on the transport in subsequent time steps.
As a consequence the model acted very much like bi=1, irrespective of the true value of bi.