Avalanching module added on the bases of Bart's fork.
bed.py has two new modules:
1. Avalanche
2. calc_grad

To get this working based on Bart's code grid definitions needed to be updated usging the gridparams.py code.

In the constants.py file the required constants are defined.

In the model.py file some small code related to the boundary conditions is updated.

Code is tested using a 2 simulation and conservation of mass was confirmed.

TODO: there seems to be some overlap in the grid definitions since the old grid definitions still exist alongside the new ones.