Implemented grain size dependent Bagnold equation (bagnold_gs)
Hoonhout & de Vries assumed that d/D = 1 in their work. In this implementation of the Bagnold equation, d/D is based on a reference grain size, D, of 250 micrometer and on the grain size, d, of each individual grain size fraction for which equilibirum transport is calculated. This implementation follows the work of Bagnold more closely, but this way of applying the Bagnold equation has never been validated for multi-fraction sediment transport calculations.