Added grainspeed functionality
When calling 'duran_full' under the grainspeed parameter, now the grainspeed equation (eq 1.61 in thesis Duran) is completely solved. This results in better outcomes for steep slopes. Implications for earlier simulations, e.g. barchans, parabolics is expected to be limited, but for more natural situations (mainly blowouts) this has a large (positive) impact on the outcomes.