Major update V3 (#181)
* commit stash

* added simulations and changed code

* Delete run1.log

* work on solvers and new examples

- Updated steady state solver
- Updated euler explicit solver
- Nearest neighbor interpolation introduced to wind series interpolation. This avoids awkward values of wind directions.

Both solvers eliminate the iteration of the linear system of equations. Performance of these solvers looks very good so far.

No compatibility with spatially varying wind fields yet. Solvers only work when winds are located in 1 quadrant.

* Update sweeping methods

Sweeping method 2 has been introduced to facilitate spatially varying wind field. The solver seems not mass-conserving yet.

* Update grainspeed duran

Duran grainspeed improved, such that us and un are 0 when ustar < uth

* Update threshold mask

Fixed usage of threshold mask. 2D mask can be given as input and will be used to mask uth for each fraction over a loop

* Added vver-mask

The addition of a vegetation-growth mask allows to disable vegetation growth at certain parts of the domain (for instance where vegetation species are non-burial resistant)

* Updated sweeping function

1. Included grain speed us & un from both upwind and current cell
2. Grain speeds determined at cell edge instead of center (now fluxes for mass conservation (and asymmetry))

* Added grainspeed functionality

When calling 'duran_full' under the grainspeed parameter, now the grainspeed equation (eq 1.61 in thesis Duran) is completely solved. This results in better outcomes for steep slopes. Implications for earlier simulations, e.g. barchans, parabolics is expected to be limited, but for more natural situations (mainly blowouts) this has a large (positive) impact on the outcomes.

* Stable sweep solver for spatially varying wind

solver could be optimized for speed
boundary conditions are still under development.

* Solver update

Most functionalities are operational but some work remains on the boundaries.

* Various improvements on the steady state solver

sweep3 is the new solver code. This is tested for continuity on all quadrants.

Sweep tests folder is added. This folder is to be deleted later.

* supply file, wind input, stopping criteria

- make facility for using supply file
- nearest interpolation of wind input
- sweep stopping criteria 1e-10 instead of 0

* Ensure water level is up-to-date with bed level after avalanching

* wet_bed_reset with TWL instead of zs

* Changes in bed.py for better names of functions in vegetation module

* Change in naming of vegetation parameters in constants.py

* Include grain velocity in visualization function

* Rotate us/un at end of model time step

* Fixes in wind.py for 1D separation bubble

* Vegetation updated according newest naming conventions

Vegetation.py updated, but not a definite version. Processes are missing and implementation is lacking. Waiting for a better description, before a clean and "final" implementation

* Update test_model.py

* circular nearest interpolation patch

* Update wind.py

Allow messy input files

* Landform simulation configurations included

* Changed example directory vanWesten2024_inreview to vanWesten2024

---------

Co-authored-by: bartvanwesten <bart.v.w@live.nl>