Add mixing option
Before, after hydrodynamic mixing, the bed composition of the mixed layers was reset to the average of those layers.

On shorter term scale, this approach seems to work, but after longer simulation periods, the approach appears to result in an underestimation of fine particles emerging in the top layer. If armouring occurs for a long period, the average bed composition of all layers starts to become more coarse. After several years, during mixing no small particles were emerging to the top layer anymore.

To solve this, an optional mixing method could be implemented that does not reset the mixed bed to the current average of the layers, but to the original bed composition.

method_mixing = layer_average (default)

OR

method_mixing = reset_initial (new)